Fix code spelling mistake.
