+++ pkg/merge/aggregator.go
@@ -15,39 +15,14 @@
  * limitations under the License.
  */
 
-package aggregater
+package merge
 
 import (
 	gxbig "github.com/dubbogo/gost/math/big"
 )
 
-type MaxAggregater struct {
-	//max  decimal.Decimal
-	max  *gxbig.Decimal
-	init bool
-}
+type Aggregator interface {
+	Aggregate(values []interface{})
 
-func (s *MaxAggregater) Aggregate(values []interface{}) {
-	if len(values) == 0 {
-		return
-	}
-
-	val, err := parseDecimal2(values[0])
-	if err != nil {
-		panic(err)
-	}
-
-	if !s.init {
-		s.max = val
-		s.init = true
-	} else if s.max.Compare(val) < 0 {
-		s.max = val
-	}
-}
-
-func (s *MaxAggregater) GetResult() (*gxbig.Decimal, bool) {
-	if s.init {
-		return s.max, true
-	}
-	return nil, false
+	GetResult() (*gxbig.Decimal, bool)
 }
