+++ pkg/merge/aggregator/add_aggregator_test.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package aggregater
+package aggregator
 
 import (
 	"testing"
@@ -27,7 +27,7 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestAvgAggregater(t *testing.T) {
+func TestAddAggregator(t *testing.T) {
 	params := []struct {
 		nums   [][]interface{}
 		result *gxbig.Decimal
@@ -35,42 +35,37 @@ func TestAvgAggregater(t *testing.T) {
 	}{
 		{
 			nums: [][]interface{}{
-				{1, 1}, {2, 2}, {7, 3}, {1.6, 2}, {10, 5}, {12.12, 3.86},
+				{1}, {2}, {3}, {1.6}, {10}, {12.12},
 			},
-			result: newDecFromFloat(2),
+			result: newDecFromFloat(29.72),
 			valid:  true,
 		},
 		{
 			nums: [][]interface{}{
-				{10, 2},
+				{0.1}, {0.2}, {10.00}, {20.10}, {},
 			},
-			result: newDecFromFloat(5),
+			result: newDecFromFloat(30.4),
 			valid:  true,
-		},
-		{
-			nums: [][]interface{}{
-				{}, {123},
-			},
-			result: nil,
-			valid:  false,
 		},
 	}
 
 	for _, param := range params {
-		addAggr := AvgAggregater{}
+		addAggr := AddAggregator{}
 		for _, agg := range param.nums {
 			addAggr.Aggregate(agg)
 		}
 		resp, err := addAggr.GetResult()
-		if param.result == nil {
-			assert.Nil(t, resp)
-		} else {
-			f1, err1 := param.result.ToFloat64()
-			f2, err2 := resp.ToFloat64()
-			assert.Nil(t, err1)
-			assert.Nil(t, err2)
-			assert.EqualValues(t, f1, f2)
-			assert.Equal(t, err, param.valid)
-		}
+		f1, err1 := param.result.ToFloat64()
+		f2, err2 := resp.ToFloat64()
+		assert.Nil(t, err1)
+		assert.Nil(t, err2)
+		assert.EqualValues(t, f1, f2)
+		assert.Equal(t, err, param.valid)
 	}
+}
+
+func newDecFromFloat(f float64) *gxbig.Decimal {
+	dec := &gxbig.Decimal{}
+	dec.FromFloat64(f)
+	return dec
 }
